requirements.txt
pygame==2.6.1
#numpy>=1.21.0
pillow